Use libappindicator for tray support rather than gtk tray status
GTK's tray status icon seemed to be broken for me under Fedora 32.
Given that appindicator is overall better, I implemented a quick
and dirty menu replacement using libappindicator. I'm not sure
that this menu supports images in it, but it's good enough for me
for the time being.

Beyond this, StatusNotifierItem may be the best way to go in the
long term.